styling + extracting components
