+++ client/src/components/games/GameDetails.js
@@ -4,10 +4,8 @@ import {Redirect} from 'react-router-dom'
 import {getGames, joinGame, updateGame, syncGame, updatePosition} from '../../actions/games'
 import {getUsers} from '../../actions/users'
 import {userId} from '../../jwt'
-import Paper from 'material-ui/Paper'
 import './GameDetails.css'
 import BoardWrapper from './board/BoardWrapper'
-import Typography from '../../../node_modules/material-ui/Typography/Typography'
 
 class GameDetails extends PureComponent {
 
@@ -79,7 +77,6 @@ class GameDetails extends PureComponent {
       .map(p => p.userId)[0]
 
     return (<div className="GameDetail-div"  tabIndex="0">
-      <Typography>
       <h1>Game #{game.id}</h1>
 
       <p>Status: {game.status}</p>
@@ -107,7 +104,6 @@ class GameDetails extends PureComponent {
         game.status !== 'pending' &&
         <BoardWrapper board={game.board} playerNumber={player.player} coordinates_p1={game.coordinates_p1} coordinates_p2={game.coordinates_p2} onKeyPressed={this.onKeyPressed} game={game}/>
       }
-      </Typography>
     </div>)
   }
 }
