+++ client/src/components/games/board/Board.css
@@ -12,7 +12,7 @@
 .board-tile {
   box-sizing: border-box;
   display: inline-block;
-  margin: 0 -1px -4px 0;
+  margin: 0 -1px -1px 0;
   height: 51px;
   width: 51px;
   border: 1px solid #c3c3c3;
