+++ client/src/index.css
@@ -2,5 +2,5 @@
 body {
   margin: 0;
   padding: 0;
-  font-family: sans-serif;
+  font-family: 'Orbitron', sans-serif;
 }
